pyproject.toml
[project]
name = "time2guess-bi"
version = "0.1.0"
description = "BI for the Time2Guess project"
authors = [
    {name = "Valentin Menoret",email = "109912731+Spookywy@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "pandas-stubs (>=2.3.2.250926,<3.0.0.0)",
    "plotly (>=6.3.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1.0"
nbformat = "^5.10.4"

